// src/lib/types/steel/StabilityCalc.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CheckStatus } from "./CheckStatus";
import type { ForceStateSnapshot } from "./ForceStateSnapshot";
import type { NamedValue } from "./NamedValue";
import type { UnityCheck } from "./UnityCheck";

/**
 * Mirror of ResistanceCalc but for stability checks.
 */
export type StabilityCalc = { id: string, title: string, article: string, force_state: ForceStateSnapshot, formula_latex: string, variables: Array<NamedValue>, intermediate_values: Array<NamedValue>, value: number, unit: string, uc: UnityCheck | null, status: CheckStatus, notes: Array<string>, };